requirements.txt
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.60.1
greenlet==3.2.4
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psycopg==3.2.13
psycopg-binary==3.2.13
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2